requirements.txt
Django==2.0.2
django-ckeditor==5.3.0
django-google-maps==0.10.1
django-js-asset==1.0.0
django-modeltranslation==0.12.1
django-solo==1.1.3
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
